#S07P31B102-367 fix. 타입 변경, 민팅x 이미지 경로 수정

+++ Back/selly-trade-service/src/main/java/com/nct/sellytradeservice/model/repository/TradeLogRepository.java
@@ -29,5 +29,5 @@ public interface TradeLogRepository extends JpaRepository<TradeLog, Long> {
   HistoryDto selectDateHistory(@Param("historyDate") String date, @Param("articleId") Long articleId);
 
   @Query(value = "SELECT AVG(tradePrice) AS avgPrice FROM TradeLog WHERE articleId = :articleId", nativeQuery = true)
-  Float selectTotalAvg(@Param("articleId") Long articleId);
+  Double selectTotalAvg(@Param("articleId") Long articleId);
 }
